requirements.txt
beautifulsoup4==4.9.3
PyYAML==5.4.1
soupsieve==2.1
